1_config/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
reportlab>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pyautogui>=0.9.54
python-telegram-bot>=20.0
faster-whisper>=0.10.0
playwright>=1.43.0
tavily-python>=0.5.0
